[FIX] l10n_vn_edi_viettel: signs on credit notes
Description of the issue/feature this PR addresses:

- Viettel SInvoice will automatically change the signs of the
amount values (except for unit price) we sent for credit notes
(on accounts with data validation step). So the "right" values
will appear in the portal. Currently (before the commit) the
values have the wrong sign in the portal (because they are
changed but we sent the right ones already). We need to retain
the original signs of those values when submitting them to SInvoice.

Current behavior before PR:

- All the amount values, including itemTotalAmountWithoutTax,
taxAmount, itemTotalAmountAfterDiscount and itemTotalAmountWithTax
are sent to Viettel SInvoice in negative values for credit note.
Viettel SInvoice changes the sign of these values to positive
(which is not correct for credit note).

Desired behavior after PR is merged:

- Only the sign of unitPrice is changed when we send the credit note
to SInvoice.

Note:
- The signs will not automatically be changed for accounts with data
validation step removed. So this commit will break credit notes for
them (since the signs are not automatically adjusted but our values
have the wrong signs).

closes odoo/odoo#218962

X-original-commit: 044833804c9e10048ec10d7e982c98c4f33f4cf4
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>