[FIX] point_of_sale: prevent missing error when closing session
Before this commit, if a POS session was open on two devices and closed
on one, a new session would be automatically created and both devices
would switch to it. However, if the "Backend" button was clicked on one
device—removing the session—then clicking "Backend" on the second device
would trigger a missing record error, since the session was already
deleted.

opw-4709064

closes odoo/odoo#205176

Signed-off-by: Adrien Guilliams (adgu) <adgu@odoo.com>