[FIX] point_of_sale: prevent cancelled orders from being marked as paid
Before this commit, if the state of an order was set to 'cancel' and
then the order was synced to the server, it would incorrectly be saved
as 'paid'. This could happen in a POS restaurant environment, for
example, when releasing a table. As a result, empty orders would appear
as paid, which is incorrect.

This behavior was particularly problematic in certain localizations
where such empty, paid orders lacked required fiscal data, leading to
errors during session closing.

opw-4714877

closes odoo/odoo#205334

Signed-off-by: David Monnom (moda) <moda@odoo.com>