[FIX] mail: soft_reload when applying an activity plan
Scheduling a custom activity plan on a single record should do a `soft_reload`
of the current view in order to avoid breadcrumbs pollution.

The most frequent use case for that action is scheduling a plan from the chatter
of the form view of a record, and retriggering an unnamed default form
act_window would only restrict possible operations for the user.

After this commit, the current view will be reloaded and the user will most
probably be able to see its enabled plan in the chatter.

task-4525830

closes odoo/odoo#210360

X-original-commit: f8880b109deea9dd300d1aa5ab82c6f58cd93c35
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>